requirements.txt
psycopg2==2.9.5
PyMySQL==1.0.2
python-dotenv==1.0.0
